Button Home and Explore can be clicked to change page Explore and page Home in communities
