+++ app/src/main/java/vn/edu/usth/x/community/Cm_HomeFragment.java
@@ -7,53 +7,7 @@ import android.view.View;
 import android.view.ViewGroup;
 import vn.edu.usth.x.R;
 
-/**
- * A simple {@link Fragment} subclass.
- * Use the {@link Cm_HomeFragment#newInstance} factory method to
- * create an instance of this fragment.
- *
- */
 public class Cm_HomeFragment extends Fragment {
-
-    // TODO: Rename parameter arguments, choose names that match
-    // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-    private static final String ARG_PARAM1 = "param1";
-    private static final String ARG_PARAM2 = "param2";
-
-    // TODO: Rename and change types of parameters
-    private String mParam1;
-    private String mParam2;
-
-    /**
-     * Use this factory method to create a new instance of
-     * this fragment using the provided parameters.
-     *
-     * @param param1 Parameter 1.
-     * @param param2 Parameter 2.
-     * @return A new instance of fragment Cm_HomeFragment.
-     */
-    // TODO: Rename and change types and number of parameters
-    public static Cm_HomeFragment newInstance(String param1, String param2) {
-        Cm_HomeFragment fragment = new Cm_HomeFragment();
-        Bundle args = new Bundle();
-        args.putString(ARG_PARAM1, param1);
-        args.putString(ARG_PARAM2, param2);
-        fragment.setArguments(args);
-        return fragment;
-    }
-
-    public Cm_HomeFragment() {
-        // Required empty public constructor
-    }
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        if (getArguments() != null) {
-            mParam1 = getArguments().getString(ARG_PARAM1);
-            mParam2 = getArguments().getString(ARG_PARAM2);
-        }
-    }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
